refactored utils directory
