+++ ecco_pipeline/utils/pipeline_utils/config_validator.py
@@ -2,6 +2,7 @@ from glob import glob
 import logging
 from jsonschema import validate, ValidationError
 import yaml
+import os
 
 logger = logging.getLogger('pipeline')
 
@@ -13,14 +14,11 @@ def validate_configs():
         schema = yaml.load(f, Loader=yaml.Loader)
 
     for filepath in configs:
-        logger.debug(f'Validating {filepath.split("/")[-1]}')
+        logger.debug(f'Validating {os.path.basename(filepath)}')
         with open(filepath, 'r') as f:
             config = yaml.load(f, Loader= yaml.Loader)
             try:
                 validate(config, schema)
             except ValidationError as e:
                 logger.error(e)
-                raise ValidationError(f'{filepath.split("/")[-1]} is malformed.')
-
-if __name__ == '__main__':
-    validate_configs()
+                raise ValidationError(f'{os.path.basename(filepath)} is malformed.')
