+++ ecco_pipeline/utils/pipeline_utils/log_config.py
