+++ ecco_pipeline/utils/processing_utils/llc_array_conversion.py
